--- service/controller/chart/v1/resource/tiller/create.go
@@ -1,20 +1,0 @@
-package tiller
-
-import (
-	"context"
-
-	"github.com/giantswarm/microerror"
-)
-
-func (r *Resource) EnsureCreated(ctx context.Context, obj interface{}) error {
-	r.logger.LogCtx(ctx, "level", "debug", "message", "ensuring tiller is installed")
-
-	err := r.helmClient.EnsureTillerInstalled(ctx)
-	if err != nil {
-		return microerror.Mask(err)
-	}
-
-	r.logger.LogCtx(ctx, "level", "debug", "message", "ensured tiller is installed")
-
-	return nil
-}
