--- service/controller/chart/v1/resource/tiller/delete.go
@@ -1,10 +1,0 @@
-package tiller
-
-import (
-	"context"
-)
-
-// EnsureDeleted is not implemented for the tiller resource.
-func (r *Resource) EnsureDeleted(ctx context.Context, obj interface{}) error {
-	return nil
-}
